+++ src/ui/mod.rs
@@ -2,5 +2,8 @@
 
 mod constants;
 mod app;
+mod styles;
+mod search;
+mod favorite;
 
 pub use app::DictionaryApp;
